requirements.txt
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.0.1
openpyxl==3.1.2
numpy==1.24.3
matplotlib==3.7.2